+++ plugin-kit/tapdata-pdk-cli/src/main/java/io/tapdata/pdk/tdd/tests/v2/TableCountTest.java
@@ -1,12 +1,8 @@
 package io.tapdata.pdk.tdd.tests.v2;
 
-import io.tapdata.entity.schema.TapTable;
 import io.tapdata.pdk.apis.context.TapConnectorContext;
-import io.tapdata.pdk.apis.functions.PDKMethod;
 import io.tapdata.pdk.apis.functions.connector.target.DropTableFunction;
 import io.tapdata.pdk.cli.commands.TapSummary;
-import io.tapdata.pdk.core.api.PDKIntegration;
-import io.tapdata.pdk.core.monitor.PDKInvocationMonitor;
 import io.tapdata.pdk.tdd.core.PDKTestBase;
 import io.tapdata.pdk.tdd.core.SupportFunction;
 import io.tapdata.pdk.tdd.tests.support.TapAssert;
@@ -17,19 +13,23 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
 import java.lang.reflect.Method;
-import java.util.Arrays;
+import java.math.BigDecimal;
 import java.util.List;
+import java.util.UUID;
 
 import static io.tapdata.entity.simplify.TapSimplify.*;
 import static io.tapdata.entity.utils.JavaTypesToTapTypes.JAVA_Long;
+import static io.tapdata.entity.utils.JavaTypesToTapTypes.JAVA_String;
 
 @DisplayName("tableCount.test")//tableCount表数量， 必测方法
 @TapGo(sort = 5)
 public class TableCountTest extends PDKTestBase {
-    protected TapTable targetTable = table(testTableId)
-            .add(field("id", JAVA_Long).isPrimaryKey(true).primaryKeyPos(1))
-            .add(field("name", "STRING"))
-            .add(field("text", "STRING"));
+    private void setTable(){
+        this.targetTable = table(UUID.randomUUID().toString())
+                .add(field("id", JAVA_Long).isPrimaryKey(true).primaryKeyPos(1).tapType(tapNumber().maxValue(BigDecimal.valueOf(Long.MAX_VALUE)).minValue(BigDecimal.valueOf(Long.MIN_VALUE))))
+                .add(field("name", JAVA_String).tapType(tapString().bytes(50L)))
+                .add(field("text", JAVA_String).tapType(tapString().bytes(50L)));
+    }
     @DisplayName("tableCount.findTableCount")//用例1， 查询表数量
     @Test
     @TapTestCase(sort = 1)
@@ -39,6 +39,7 @@ public class TableCountTest extends PDKTestBase {
     void findTableCount(){
         super.consumeQualifiedTapNodeInfo(nodeInfo -> {
             TestNode prepare = this.prepare(nodeInfo);
+            //setTable();
             try {
                 super.connectorOnStart(prepare);
                 prepare.recordEventExecute().testCase(super.getMethod("findTableCount"));
@@ -58,16 +59,16 @@ public class TableCountTest extends PDKTestBase {
             tableCount = prepare.connectorNode().getConnector().tableCount(connectorContext);
         }catch (Exception e){
             TapAssert.asserts(()->
-                    Assertions.fail(TapSummary.format("tableCount.findTableCount.errorFun",e.getMessage()))
+                Assertions.fail(TapSummary.format("tableCount.findTableCount.errorFun",e.getMessage()))
             ).acceptAsError(testCase, null);
         }
         int tableCountFinal = tableCount;
         TapAssert.asserts(()->
-                Assertions.assertTrue(tableCountFinal > 0,
-                        TapSummary.format("tableCount.findTableCount.error",tableCountFinal)
-                )
+            Assertions.assertTrue(tableCountFinal > 0,
+                TapSummary.format("tableCount.findTableCount.error",tableCountFinal)
+            )
         ).acceptAsError(testCase,
-                TapSummary.format("tableCount.findTableCount.succeed",tableCountFinal)
+            TapSummary.format("tableCount.findTableCount.succeed",tableCountFinal)
         );
         return tableCount;
     }
@@ -84,10 +85,11 @@ public class TableCountTest extends PDKTestBase {
     void findTableCountAfterNewTable(){
         super.consumeQualifiedTapNodeInfo(nodeInfo -> {
             TestNode prepare = this.prepare(nodeInfo);
+            setTable();
             RecordEventExecute execute = prepare.recordEventExecute();
+            boolean hasCreatedTable = false;
             try {
                 super.connectorOnStart(prepare);
-
                 Method testCase = super.getMethod("findTableCountAfterNewTable");
                 execute.testCase(testCase);
 
@@ -95,24 +97,34 @@ public class TableCountTest extends PDKTestBase {
                 int tableCountNewTableAgo = tableCount(prepare);
 
                 //调用CreateTableFunction新建一张表，
-                if (!this.createTable(prepare)) return;
+                if (!( hasCreatedTable = this.createTable(prepare) ) ) return;
 
                 //再调用tableCount方法获得表数量，
                 int tableCountNewTableAfter = tableCount(prepare);
 
-                //比之前的数量加1就是正确的 @TODO 实际场景下不一定数量高度符合预期
+                //比之前的数量加1就是正确的 | 实际场景下不一定数量高度符合预期
                 TapAssert.asserts(()->
-                        Assertions.assertTrue(tableCountNewTableAgo+1 == tableCountNewTableAfter,
-                                TapSummary.format("tableCount.findTableCountAfterNewTable.afterNewTable.error",tableCountNewTableAgo,tableCountNewTableAfter)
+                    Assertions.assertEquals(
+                        tableCountNewTableAfter,
+                        tableCountNewTableAgo + 1,
+                        TapSummary.format(
+                            "tableCount.findTableCountAfterNewTable.afterNewTable.error",
+                            tableCountNewTableAgo,
+                            tableCountNewTableAfter
                         )
-                ).acceptAsWarn(testCase,
-                        TapSummary.format("tableCount.findTableCountAfterNewTable.afterNewTable.succeed",tableCountNewTableAgo,tableCountNewTableAfter)
+                    )
+                ).acceptAsWarn(
+                    testCase,
+                    TapSummary.format(
+                        "tableCount.findTableCountAfterNewTable.afterNewTable.succeed",
+                        tableCountNewTableAgo,
+                        tableCountNewTableAfter
+                    )
                 );
-
             }catch (Throwable e) {
                 throw new RuntimeException(e);
             }finally {
-                execute.dropTable();
+                if (hasCreatedTable) execute.dropTable();
                 super.connectorOnStop(prepare);
             }
         });
@@ -120,7 +132,7 @@ public class TableCountTest extends PDKTestBase {
 
     public static List<SupportFunction> testFunctions() {
         return list(
-                support(DropTableFunction.class, TapSummary.format(inNeedFunFormat,"DropTableFunction"))
+//                support(DropTableFunction.class, TapSummary.format(inNeedFunFormat,"DropTableFunction"))
         );
     }
 }
